moved custom SelectSearch into its own file
